+++ core/renderer/css/css_property_unittest.cc
@@ -15,7 +15,7 @@ namespace test {
 
 TEST(CSSProperty, GetComputeStyleMap) {
   auto map = CSSProperty::GetComputeStyleMap();
-  EXPECT_EQ(map.size(), 217);
+  EXPECT_EQ(map.size(), 220);
 }
 
 TEST(CSSProperty, GetPropertyID) {