+++ core/style/text_attributes.h
@@ -9,6 +9,7 @@
 #include <vector>
 
 #include "base/include/flex_optional.h"
+#include "base/include/value/array.h"
 #include "base/include/value/base_value.h"
 #include "base/include/vector.h"
 #include "core/renderer/starlight/style/css_type.h"
@@ -91,32 +92,59 @@ class TextAttributes {
   bool underline_decoration{DefaultComputedStyle::DEFAULT_BOOLEAN};
   bool line_through_decoration{DefaultComputedStyle::DEFAULT_BOOLEAN};
   bool is_auto_font_size{DefaultComputedStyle::DEFAULT_AUTO_FONT_SIZE};
+  fml::RefPtr<lepus::CArray> font_variation_settings{nullptr};
+  fml::RefPtr<lepus::CArray> font_feature_settings{nullptr};
+  starlight::FontOpticalSizingType font_optical_sizing{
+      DefaultComputedStyle::DEFAULT_FONT_OPTICAL_SIZING};
 
   void Reset() {}
 
   bool operator==(const TextAttributes& rhs) const {
-    return std::tie(font_size, color, decoration_color, white_space,
-                    text_overflow, font_weight, font_style, font_family,
-                    computed_line_height, line_height_factor,
-                    enable_font_scaling, letter_spacing, line_spacing,
-                    text_shadow, text_align, word_break, underline_decoration,
-                    line_through_decoration, text_decoration_color,
-                    text_decoration_style, text_indent, is_auto_font_size,
-                    auto_font_size_min_size, auto_font_size_max_size,
-                    auto_font_size_step_granularity,
-                    auto_font_size_preset_sizes, hyphens) ==
-           std::tie(rhs.font_size, rhs.color, rhs.decoration_color,
-                    rhs.white_space, rhs.text_overflow, rhs.font_weight,
-                    rhs.font_style, rhs.font_family, rhs.computed_line_height,
-                    rhs.line_height_factor, rhs.enable_font_scaling,
-                    rhs.letter_spacing, rhs.line_spacing, rhs.text_shadow,
-                    rhs.text_align, rhs.word_break, rhs.underline_decoration,
-                    rhs.line_through_decoration, rhs.text_decoration_color,
-                    rhs.text_decoration_style, rhs.text_indent,
-                    rhs.is_auto_font_size, rhs.auto_font_size_min_size,
-                    rhs.auto_font_size_max_size,
-                    rhs.auto_font_size_step_granularity,
-                    rhs.auto_font_size_preset_sizes, hyphens);
+    bool base_equal =
+        std::tie(font_size, color, decoration_color, white_space, text_overflow,
+                 font_weight, font_style, font_family, computed_line_height,
+                 line_height_factor, enable_font_scaling, letter_spacing,
+                 line_spacing, text_shadow, text_align, word_break,
+                 underline_decoration, line_through_decoration,
+                 text_decoration_color, text_decoration_style, text_indent,
+                 is_auto_font_size, auto_font_size_min_size,
+                 auto_font_size_max_size, auto_font_size_step_granularity,
+                 auto_font_size_preset_sizes, hyphens, font_optical_sizing) ==
+        std::tie(
+            rhs.font_size, rhs.color, rhs.decoration_color, rhs.white_space,
+            rhs.text_overflow, rhs.font_weight, rhs.font_style, rhs.font_family,
+            rhs.computed_line_height, rhs.line_height_factor,
+            rhs.enable_font_scaling, rhs.letter_spacing, rhs.line_spacing,
+            rhs.text_shadow, rhs.text_align, rhs.word_break,
+            rhs.underline_decoration, rhs.line_through_decoration,
+            rhs.text_decoration_color, rhs.text_decoration_style,
+            rhs.text_indent, rhs.is_auto_font_size, rhs.auto_font_size_min_size,
+            rhs.auto_font_size_max_size, rhs.auto_font_size_step_granularity,
+            rhs.auto_font_size_preset_sizes, rhs.hyphens,
+            rhs.font_optical_sizing);
+    if (!base_equal) {
+      return false;
+    }
+
+    if (font_variation_settings == rhs.font_variation_settings) {
+    } else if (font_variation_settings && rhs.font_variation_settings) {
+      if (*font_variation_settings != *rhs.font_variation_settings) {
+        return false;
+      }
+    } else {
+      return false;
+    }
+
+    if (font_feature_settings == rhs.font_feature_settings) {
+    } else if (font_feature_settings && rhs.font_feature_settings) {
+      if (*font_feature_settings != *rhs.font_feature_settings) {
+        return false;
+      }
+    } else {
+      return false;
+    }
+
+    return true;
   }
 
   bool operator!=(const TextAttributes& rhs) const { return !(*this == rhs); }