+++ tools/css_generator/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.6
+- Add `font-variation-settings`,`font-feature-settings` and `font-optical-sizing` CSS properties.
+
 ## 0.0.5
 
 - `vertical-align` property is `layout_wanted`, text need